pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-triage-env"
version = "1.0.0"
description = "OpenEnv environment for Emergency Department triage simulation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Team Unfazed"},
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
include = ["clinical_triage_env*"]
